GET, POST, DELETE

+++ src/services/shop/shop-router.ts
@@ -1,6 +1,10 @@
+import crypto from "crypto";
+import Config from "../../config.js";
 import { Request, Response, Router } from "express";
 import { NextFunction } from "express-serve-static-core";
 import { weakJwtVerification, strongJwtVerification } from "../../middleware/verify-jwt.js";
+import { hasAdminPerms } from "../auth/auth-lib.js";
+import { JwtPayload } from "../auth/auth-models.js";
 import { DeleteResult } from "mongodb";
 import { StatusCode } from "status-code-enum";
 import { RouterError } from "../../middleware/error-handler.js";
@@ -36,7 +40,6 @@ import Models from "../../database/models.js";
 
 const shopRouter: Router = Router();
 shopRouter.get("/", weakJwtVerification, async (_: Request, res: Response, next: NextFunction) => {
-    console.log("nesdkjfnjksfksdf");
     try {
         const shopItems: ShopItemFormat[] = await Models.ShopItem.find();
         const shopQuantities: QuantityFormat[] = await Models.ShopQuantity.find();
@@ -69,9 +72,90 @@ shopRouter.get("/", weakJwtVerification, async (_: Request, res: Response, next:
 });
 
 /**
+ * @api {post} /shop/item post /shop/item
+ * @apiGroup Shop
+ * @apiDescription Insert a new item into the shop.
+ * 
+ * @apiHeader {String} Authorization User's JWT Token with admin permissions.
+ * 
+ * @apiBody {Json} item The item details to be created.
+ * @apiParamExample {Json} Request Body Example for an Item:
+ * {
+ *      "name": "HackIllinois Branded Hoodie",
+ *      "price": 15,
+ *      "isRaffle": true,
+ *      "quantity": 1
+ * }
+ *
+ * @apiSuccess (200: Success) {Json} items The items details.
+ * @apiSuccessExample Example Success Response
+ * HTTP/1.1 200 OK
+ * {
+ *      "itemId": "item01",
+ *      "name": "HackIllinois Branded Hoodie",
+ *      "price": 15,
+ *      "isRaffle": true,
+ *      "quantity": 1
+ * }
+ *
+ * @apiUse strongVerifyErrors
+ * @apiError (400: Bad Request) {String} ExtraIdProvided Invalid item parameters provided.
+ * @apiError (400: Bad Request) {String} ItemAlreadyExists Item already exists in shop.
+ * @apiError (403: Forbidden) {String} InvalidPermission User does not have admin permissions.
+ * @apiError (500: Internal Server Error) {String} InternalError An error occurred on the server.
+ * */
+shopRouter.post("/item", strongJwtVerification, async (req: Request, res: Response, next: NextFunction) => {
+    const payload: JwtPayload = res.locals.payload as JwtPayload;
+
+    // Check if the token has admin permissions
+    if (!hasAdminPerms(payload)) {
+        return next(new RouterError(StatusCode.ClientErrorForbidden, "InvalidPermission"));
+    }
+
+    if (req.body.itemId) {
+        return next(new RouterError(StatusCode.ClientErrorBadRequest, "ExtraIdProvided", { extraItemId: req.body.itemId }));
+    }
+
+    const itemId = "item" + parseInt(crypto.randomBytes(Config.SHOP_BYTES_GEN).toString("hex"),16);
+
+    const metaItem: ItemFormat = req.body as ItemFormat;
+    metaItem.itemId = itemId;
+
+    const shopItem: ShopItemFormat = {
+        itemId: itemId,
+        name: req.body.name,
+        price: req.body.price,
+        isRaffle: req.body.isRaffle,
+    };
+
+    const itemQuantity: QuantityFormat = {
+        itemId: itemId,
+        quantity: req.body.quantity
+    };
+
+    // Ensure that user doesn't already exist before creating
+    const item1: ShopItemFormat | null = await Models.ShopItem.findOne({ itemId: shopItem.itemId });
+    const item2: QuantityFormat | null = await Models.ShopQuantity.findOne({ itemId: itemQuantity.itemId });
+    if (item1 || item2) {
+        return next(new RouterError(StatusCode.ClientErrorBadRequest, "ItemAlreadyExists"));
+    }
+
+    try {
+        await Models.ShopItem.create(shopItem);
+        await Models.ShopQuantity.create(itemQuantity);
+        return res.status(StatusCode.SuccessOK).send(metaItem);
+    } catch (error) {
+        console.error(error);
+        return next(new RouterError(StatusCode.ClientErrorBadRequest, "InvalidParams"));
+    }
+});
+
+/**
  * @api {delete} /shop/item/:ITEMID DELETE /shop/item/:ITEMID
  * @apiGroup Shop
  * @apiDescription Delete the a specific item in the point shop based itemId.
+ * 
+ * @apiHeader {String} Authorization User's JWT Token with admin permissions.
  *
  * @apiSuccess (200: Success) {Json} success Indicates successful deletion of the user's profile.
  * @apiSuccessExample Example Success Response:
@@ -80,16 +164,20 @@ shopRouter.get("/", weakJwtVerification, async (_: Request, res: Response, next:
  *    "success": true
  * }
  * 
- * @apiError (500: Internal Error) {String} InternalError An internal server error occurred.
- * @apiErrorExample Example Error Response (InternalError):
- *     HTTP/1.1 500 Internal Server Error
- *     {"error": "InternalError"}
+ * @apiUse strongVerifyErrors
+ * @apiError (403: NotFound) {String} ItemNotFound Item not found in the shop.
+ * @apiError (403: Forbidden) {String} InvalidPermission User does not have admin permissions.
+ * @apiError (500: Internal Server Error) {String} InternalError An error occurred on the server.
  */
-
 shopRouter.delete("/item/:ITEMID", strongJwtVerification, async (req: Request, res: Response, next: NextFunction) => {
-    console.log("89");
+    const payload: JwtPayload = res.locals.payload as JwtPayload;
+
+    // Check if the token has admin permissions
+    if (!hasAdminPerms(payload)) {
+        return next(new RouterError(StatusCode.ClientErrorForbidden, "InvalidPermission"));
+    }
+
     const targetItem: string | undefined = req.params.ITEMID as string;
-    console.log("91");
 
     const shopItemDeleteResponse: DeleteResult = await Models.ShopItem.deleteOne({ itemId: targetItem });
     const shopQuantityDeleteResponse: DeleteResult = await Models.ShopQuantity.deleteOne({ itemId: targetItem });
@@ -99,7 +187,5 @@ shopRouter.delete("/item/:ITEMID", strongJwtVerification, async (req: Request, r
     }
     return res.status(StatusCode.SuccessOK).send({ success: true });
 });
-
-
 
 export default shopRouter;
